auth-service/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==0.16.0
psycopg2==2.8.4
pyjwt==1.4.2
flask-bcrypt==0.7.1
